Add an end-to-end executable test for generalized bridging.
Use the swift_bridge Clang attribute and paired _ObjectiveCBridgeable
conformance to provide bridging for a an Objective-C class to a Swift
value type. All of these code paths have been exercised already by the
generalizations for String/Array/Dictionary/Set, so there are no
compiler changes required here, but this verifies that we haven't
missed any ad hoc rules for those bridged types along the critical
path.

Checkpoint that we've implemented the bulk of rdar://problem/24659951.