Only allow _ObjectiveCBridgeable conformances in the type's defining module.
With the exception of a specific whitelist of cases where the
Foundation module defines conformances to _ObjectiveCBridgeable for
standard library types, only permit an _ObjectiveCBridgeable
conformance in the same module as the type that's conforming to the
protocol. Among other things, this prevents the optimizer from
concluding that a dynamic cast between a Swift value type and its
bridged Objective-C class type can never succeed. See

  https://github.com/apple/swift/commit/34ff1c8e6df411a821b6dc7cd031506f875fb6be

for the optimizer issue. As part of this, bring the whitelist in sync
with reality, now that the compiler enforces it.