Reapply "[ReconstructType] Fix decl lookup when there are multiple constructors"
This was mistakenly reverted in an attempt to fix bot failures.  It
turned out to be a different commit that was responsible, so reapply.
Original commit message was:

We were failing to find init() decls whenever there was more than one
candidate.

The mangled function type we get will be (args...) -> T, but the real
thing is (T.Type) -> (args...) -> T. Strip the metatype so we can match
successfully.