[build-presets] Remove explicit =1 for impl params
The `utils/build-script-impl` shellscript takes arguments in the form of `--foo=bar` or `--baz`. In the second form, the argument is automatically interpreted as `--baz=1`. There is no need to manually specify `=1`, so remove it from the build presets.

There is another reason to do this: the argument parser used by the Python `utils/build-script` is not as loose as the shellscript's. For a boolean switch such as `--baz`, the argument parser throws an exception if given a numerical value such as `--baz=1`. Therefore attempts to migrate certain parameters from the shellscript to the Python build script will encounter trouble when dealing with these explicit `--baz=1` parameters in the build presets.

This commit does not touch any of the `--foo=0` parameters. Although they should be removed, it's trickier due to the inheritance of the build presets, and so I will do so in a later commit.