[build-presets] No need to disable static stdlib
Static stdlib builds are disabled unless `--build-swift-static-stdlib` is explicitly passed in to the build scripts. As a result, `--build-swift-static-stdlib=0` is pointless--unless it's being used to counteract another `--build-swift-static-stdlib` being passed to the build script, but this is not the case in any of the uses in the build presets.

Remove the unnecessary parameters to make the build presets easier to understand.