[build-script] Link Objective-C runtime for XCTest
The XCTest CI is currently broken for OS X. Invoking `xcodebuild`, even
when setting the correct `SWIFT_EXEC`, was unable to build Foundation.
The problem was that a flag to link the Objective-C runtime was not being set
as it would be when building with a Swift toolchain.

Set the flag to get the builds passing. The XCTest CI presets should
pass on all platforms with this change.