[build-presets] Foundation built based on platform
Previously, the XCTest CI preset would attempt to build
corelibs-foundation on OS X. corelibs-foundation is not meant to be
built via the build-script on OS X, so this would cause issues.

Remove the explicit `--foundation` option, and instead leave it up to
the build-script to determine whether to build Foundation when building
XCTest.