[SIL] Move type fixups into SILBuilder::getPartialApplyResultType.
We do clever tricks to avoid having to serialize the full type of a
partial_apply instruction, but those conflicted with our clever tricks
for handling unusual return conventions. Make sure we do all type
calculations for partial_apply in one place. (Credit to Slava for
both introducing the problem and immediately seeing how to fix it.)

This should allow re-application of e6a519f and 86dcce1.