[Clang importer] Fix obscure issue with submodule protocols.
In the Clang importer, and only in submodules, the first protocol
typedef with NSObject that we try to import results in a hidden lookup
result. So, allow hidden lookups, which is benign and failure is
actually an issue of a malformed header. Gracefully handles failure.

Test case included that will reproduce the issue if AllowHidden is
disabled.