Make FSO thunks always_inline.
This forces the callsites to be rewritten by the inliner.

we have the issue that the thunk changes from the time the its created to
the time its reread to figure out what we have done to the original function

This can result anywhere from a missed opportunities to a broken SIL.

This solution solves the problem gracefully, because the thunk carries the information
on how to set up the call to the optimized functions.

I did not measure any regression with this change.