XXX SIL Optimizer: Set the [thunk] attribute on specializations XXX
The original SILFunction may be [fragile] because it came from
an imported module, or maybe it was [fragile] because it was
declared @_transparent by the user.

Furthermore, the devirtualizer can devirtualize calls inside
[fragile] functions, pointing them at  non-fragile functions.

FIXME: For resilience, we cannot devirtualize this stuff anyway.
Instead, try to not set [fragile] on clones unless needed.