AST/SILGen: Factor out "function body might be inlined in other resilience domains" logic
When i made the recent change to serialization to serialize local
functions of @_transparent functions, I forgot to make the
corresponding change to the logic for deciding whether properties
should be accessed directly or via accessors.

Also, the accessor check was not taking @inline(__always) into
account.

This patch fixes both issues and removes some duplication.